api/backend_management/requirements.txt
django~=5.2.4
djangorestframework~=3.16.0
django-cors-headers~=4.7.0
coreapi~=2.3.3
djangorestframework-simplejwt~=5.5.0
django-redis~=6.0.0
apscheduler~=3.11.0
django_apscheduler~=0.7.0
drf-yasg~=1.21.10
paramiko~=2.8.1
psycopg2-binary~=2.9.10
django-filter~=25.1
pytest-django~=4.11.1
pytest-html~=4.1.1
allure-pytest~=2.13.5